+++ jdaviz/configs/imviz/tests/test_astrowidgets_api.py
@@ -159,8 +159,9 @@ class TestCmapStretchCuts(BaseImviz_WCS_NoWCS):
             'Gray', 'Viridis', 'Plasma', 'Inferno', 'Magma', 'Purple-Blue',
             'Yellow-Green-Blue', 'Yellow-Orange-Red', 'Red-Purple', 'Blue-Green',
             'Hot', 'Red-Blue', 'Red-Yellow-Blue', 'Purple-Orange', 'Purple-Green',
+            'Rainbow', 'Seismic',
             'Reversed: Gray', 'Reversed: Viridis', 'Reversed: Plasma', 'Reversed: Inferno',
-            'Reversed: Magma', 'Reversed: Hot']
+            'Reversed: Magma', 'Reversed: Hot', 'Reversed: Rainbow']
 
     def test_invalid_colormap(self):
         with pytest.raises(ValueError, match='Invalid colormap'):